requirements.txt
aiofiles
altair
annotated-types
anyio
attrs
Brotli
certifi
charset-normalizer
click
colorama
contourpy
cycler
dnspython
email_validator
exceptiongroup
fastapi
fastapi-cli
ffmpy
filelock
fonttools
fsspec
gradio
gradio_client
h11
h2
hpack==4.0.0
httpcore
httpx
huggingface_hub
hyperframe
idna
importlib_metadata
importlib_resources
Jinja2
jsonschema
jsonschema-specifications
kiwisolver
markdown-it-py
MarkupSafe
matplotlib
mdurl
munkres==1.1.4
numpy==1.26.4
orjson==3.10.3
packaging
pandas
pillow
pkgutil_resolve_name
pydantic
pydantic_core
pydub
Pygments
pyparsing
PySocks
python-dateutil
python-multipart
pytz
PyYAML
referencing
requests
rich
rpds-py
ruff
semantic-version
setuptools
shellingham
six
sniffio
starlette
tomlkit
toolz
torch
torchvision
tqdm
typer==0.12.3
typer-slim==0.12.3
typing_extensions
tzdata
ujson
urllib3
uvicorn
websockets
wheel==0.43.0
zipp
